requirements.txt
opencv-contrib-python<4.10
face-recognition>=1.3.0
dlib>=19.22
numpy<2.0
Pillow>=9.0
twilio>=8.0
